refactor: Convert financial fields to Decimal for improved precision
- Updated the Prisma schema to change various financial fields from BigInt to Decimal, ensuring accurate handling of large values without JSON serialization issues.
- Modified repository methods and data interfaces to accommodate the new Decimal type for balances, fees, and amounts across multiple models including Account, Extrinsic, Transfer, and Validator.
- Enhanced job processing and indexing logic to utilize Decimal for financial calculations, improving overall data integrity and performance.
- Added a new script for database inspection to facilitate monitoring of data integrity post-migration.

This commit enhances the financial data handling capabilities of the application, ensuring precision and consistency across the codebase.

+++ src/services/domain/account/AccountInterfaces.ts
@@ -8,6 +8,7 @@
 
 import { Account, Extrinsic, Validator, Reward } from '@prisma/client';
 import { TransferWithRelations } from '../../../database/repositories/TransferRepository';
+import { Decimal } from '@prisma/client/runtime/library';
 
 /**
  * Account balance information from blockchain RPC
@@ -47,9 +48,9 @@ export interface AccountWithDetails extends Account {
   /** Total number of extrinsics signed by this account */
   extrinsicCount: number;
   /** Total amount transferred out by this account */
-  totalTransferred: bigint;
+  totalTransferred: Decimal;
   /** Total amount received by this account */
-  totalReceived: bigint;
+  totalReceived: Decimal;
 }
 
 /**
@@ -87,9 +88,9 @@ export interface AccountStats {
   /** Date of most recent activity */
   lastActivity: Date | null;
   /** Total amount sent in transfers */
-  totalSent: bigint;
+  totalSent: Decimal;
   /** Total amount received in transfers */
-  totalReceived: bigint;
+  totalReceived: Decimal;
 }
 
 /**
